GoogleTranslateOption conflict
GoogleTranslateOption added an extra context menu entry on the message field. To circumvent this I added a check to see if it had started, if so, it'd mean the edit button had moved a section further down thus increasing it's position as an index value (3rd (2) over 2nd (1)).

Let me know if there're any other apps that you have conflicts with, I'll be unable to try all available plugins on the planet myself.