Added "PersonalPins" as an Index conflict
https://github.com/Farcrada/Double-click-to-edit/issues/9 mentioned another plugin which causes conflicts.

Having realized this list is eventually going to become too long to just add on to (and probably going to shift depending on what plugins are active) so I'll be rewriting this in the future to be more adjusted. For now this should work, let me know if it doesn't.