Like father, like son.
The plugin is now as I had envisioned it.

There are little QoL improvements to be made (text for example), and I'm open to suggestions, simply leave a comment or open an issue. ❤️ 

The previous version was quite intrusive. Although it worked it was impeding on user experience. To combat this I had to create a tool-tip. I had recently refactored a lot of the design to make it work reliably, only to find out that it had a 34 channel limit(*). To fix this there would be little use trying not to constantly check for mouse activity, and just use tool-tips like Discord itself does.

I am quite happy with how it is now any future updates will most likely only be to adjust to Discord changes.

(*)The `onSwitch()` had a limitation for only the present channels in Discord. These are dynamically loaded in big servers (34+ channels & category headers).