Improved on the formatting and look/feel
There wasn't anything inherently wrong with it, but it was severely outdated and the text simply looked bad and out of place.
I removed duplicate elements that were getting called.
And improved role definition.
Also added a nice little effect to it.

There is but one thing that is bothering me around line 337. It feels very scuffed and I feel it should be different, but it works.