Fix category sync and voice implementation
The category sync had fallen apart and couldn't make reliable comparisons. In addition to this all roles should now show up in the same order as a result/cause of this fix.

Voice channels didn't get much love. It worked but it was horrible to look at, this has now been fixed. It should look like it belongs now and should feel a lot better to use.