Fix #96
This fix concerns One of Discord Experiment's features: `Favorite Server`. It lets you add any channel ((group) DM or a server's channel) to a favorites section for easy access.

When adding (group) DMs to favorites, the logic failed to find a guild. Which wasn't entirely unexpected seen as it's a local list but just not accounted for. This should fix that.