Change button placement logic
Previous attempt was through manipulation of the context menu as a whole.

New (or rather: old) method is (was) to directly address the button below it "Call" and add the "Join Call" above it. This way the plugin doesn't have to look around for anything and simply adds itself to it.